test: isolate span-emission tests into separate integration test file to prevent subscriber conflicts on ubuntu
