+++ autumn-harvest/src/executor.rs
@@ -534,304 +534,4 @@ mod tests {
             other => panic!("expected Completed, got {other:?}"),
         }
     }
-
-    // ── RED-phase span-emission tests ──────────────────────────────────────────
-    // These assert the ADR-0001 span contract: correct names, attribute keys from
-    // the ATTR_* constants, and replay semantics.  They drive the GREEN-phase
-    // changes in this module.
-
-    /// Minimal subscriber layer that records every span name seen during a test.
-    mod span_capture {
-        use std::sync::{Arc, Mutex};
-        use tracing::Subscriber;
-        use tracing_subscriber::Layer;
-
-        #[derive(Clone, Default)]
-        pub struct SpanNames(pub Arc<Mutex<Vec<String>>>);
-
-        impl SpanNames {
-            pub fn has(&self, name: &str) -> bool {
-                self.0.lock().unwrap().iter().any(|n| n == name)
-            }
-        }
-
-        pub struct SpanNameLayer(pub SpanNames);
-
-        impl<S: Subscriber> Layer<S> for SpanNameLayer {
-            fn on_new_span(
-                &self,
-                attrs: &tracing::span::Attributes<'_>,
-                _id: &tracing::span::Id,
-                _ctx: tracing_subscriber::layer::Context<'_, S>,
-            ) {
-                self.0
-                    .0
-                    .lock()
-                    .unwrap()
-                    .push(attrs.metadata().name().to_string());
-            }
-        }
-
-        /// Subscriber that captures per-span field values as `(field_name, value_string)`.
-        #[derive(Clone, Default)]
-        #[allow(clippy::type_complexity)]
-        pub struct SpanFields(pub Arc<Mutex<Vec<(String, Vec<(String, String)>)>>>);
-
-        impl SpanFields {
-            /// Returns all field values recorded for spans with the given name.
-            pub fn fields_for(&self, span_name: &str) -> Vec<Vec<(String, String)>> {
-                self.0
-                    .lock()
-                    .unwrap()
-                    .iter()
-                    .filter(|(n, _)| n == span_name)
-                    .map(|(_, f)| f.clone())
-                    .collect()
-            }
-        }
-
-        struct FieldRecorder(Vec<(String, String)>);
-        impl tracing::field::Visit for FieldRecorder {
-            fn record_debug(&mut self, field: &tracing::field::Field, value: &dyn std::fmt::Debug) {
-                self.0
-                    .push((field.name().to_string(), format!("{value:?}")));
-            }
-            fn record_str(&mut self, field: &tracing::field::Field, value: &str) {
-                self.0.push((field.name().to_string(), value.to_string()));
-            }
-            fn record_bool(&mut self, field: &tracing::field::Field, value: bool) {
-                self.0.push((field.name().to_string(), value.to_string()));
-            }
-        }
-
-        pub struct SpanFieldLayer(pub SpanFields);
-
-        impl<S: Subscriber> Layer<S> for SpanFieldLayer {
-            fn on_new_span(
-                &self,
-                attrs: &tracing::span::Attributes<'_>,
-                _id: &tracing::span::Id,
-                _ctx: tracing_subscriber::layer::Context<'_, S>,
-            ) {
-                let mut visitor = FieldRecorder(Vec::new());
-                attrs.record(&mut visitor);
-                self.0
-                    .0
-                    .lock()
-                    .unwrap()
-                    .push((attrs.metadata().name().to_string(), visitor.0));
-            }
-        }
-    }
-
-    // ── ADR §2.1: harvest.workflow.execute span name ──────────────────────────
-    // These are plain `#[test]` (not `#[tokio::test]`) so they can build their
-    // own Tokio runtime inside `tracing::subscriber::with_default`, which requires
-    // a synchronous closure.
-
-    /// `run_workflow` must emit a span named `harvest.workflow.execute`
-    /// (the ADR-0001 §2.1 span name).
-    ///
-    /// RED: currently emits `harvest.workflow.run` — this test fails before the fix.
-    #[test]
-    fn executor_emits_harvest_workflow_execute_span() {
-        use span_capture::{SpanNameLayer, SpanNames};
-        use tracing_subscriber::prelude::*;
-
-        let names = SpanNames::default();
-        let subscriber = tracing_subscriber::registry().with(SpanNameLayer(names.clone()));
-        let _guard = tracing::subscriber::set_default(subscriber);
-
-        let exec_id = ExecutionId::new();
-        let history = vec![WorkflowEvent::WorkflowStarted {
-            input: Value::Null,
-            timestamp: Utc::now(),
-        }];
-
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(run_workflow(exec_id, history, echo_workflow, Value::Null));
-
-        assert!(
-            names.has("harvest.workflow.execute"),
-            "expected span 'harvest.workflow.execute' but only saw: {:?}",
-            names.0.lock().unwrap()
-        );
-    }
-
-    /// `run_workflow` must NOT emit a span named `harvest.workflow.run`.
-    ///
-    /// RED: currently it does — this test fails before the rename.
-    #[test]
-    fn executor_no_longer_emits_old_harvest_workflow_run_span() {
-        use span_capture::{SpanNameLayer, SpanNames};
-        use tracing_subscriber::prelude::*;
-
-        let names = SpanNames::default();
-        let subscriber = tracing_subscriber::registry().with(SpanNameLayer(names.clone()));
-        let _guard = tracing::subscriber::set_default(subscriber);
-
-        let exec_id = ExecutionId::new();
-        let history = vec![WorkflowEvent::WorkflowStarted {
-            input: Value::Null,
-            timestamp: Utc::now(),
-        }];
-
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(run_workflow(exec_id, history, echo_workflow, Value::Null));
-
-        assert!(
-            !names.has("harvest.workflow.run"),
-            "old span 'harvest.workflow.run' must be removed; saw: {:?}",
-            names.0.lock().unwrap()
-        );
-    }
-
-    // ── ADR §2.1: harvest.execution.id attribute ──────────────────────────────
-
-    /// The `harvest.workflow.execute` span must carry `harvest.execution.id`
-    /// (i.e. `ATTR_EXECUTION_ID`), not the old `workflow.execution_id` field.
-    ///
-    /// RED: current code uses `workflow.execution_id` as the field name.
-    #[test]
-    fn executor_span_has_attr_execution_id_field() {
-        use crate::telemetry::ATTR_EXECUTION_ID;
-        use span_capture::{SpanFieldLayer, SpanFields};
-        use tracing_subscriber::prelude::*;
-
-        let fields = SpanFields::default();
-        let subscriber = tracing_subscriber::registry().with(SpanFieldLayer(fields.clone()));
-        let _guard = tracing::subscriber::set_default(subscriber);
-
-        let exec_id = ExecutionId::new();
-        let history = vec![WorkflowEvent::WorkflowStarted {
-            input: Value::Null,
-            timestamp: Utc::now(),
-        }];
-
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(run_workflow(exec_id, history, echo_workflow, Value::Null));
-
-        let span_fields = fields.fields_for("harvest.workflow.execute");
-        assert!(
-            !span_fields.is_empty(),
-            "no 'harvest.workflow.execute' span was emitted"
-        );
-        let any_has_exec_id = span_fields
-            .iter()
-            .any(|field_set| field_set.iter().any(|(name, _)| name == ATTR_EXECUTION_ID));
-        assert!(
-            any_has_exec_id,
-            "span must carry field '{ATTR_EXECUTION_ID}'; saw fields: {span_fields:?}"
-        );
-    }
-
-    // ── ADR §2.1: harvest.replay attribute on replay spans ────────────────────
-
-    /// `run_workflow_strict` is the replay path: it must emit `harvest.workflow.execute`
-    /// with `harvest.replay = true`.
-    ///
-    /// RED: currently emits `harvest.workflow.run_strict` with no replay attribute.
-    #[test]
-    fn replay_executor_emits_harvest_workflow_execute_with_replay_true() {
-        use crate::context::empty_shared_state;
-        use crate::telemetry::ATTR_REPLAY;
-        use span_capture::{SpanFieldLayer, SpanFields};
-        use tracing_subscriber::prelude::*;
-
-        let fields = SpanFields::default();
-        let subscriber = tracing_subscriber::registry().with(SpanFieldLayer(fields.clone()));
-        let _guard = tracing::subscriber::set_default(subscriber);
-
-        let exec_id = ExecutionId::new();
-        // Provide a complete single-event history so the strict executor
-        // completes cleanly without non-determinism errors.
-        let history = vec![WorkflowEvent::WorkflowStarted {
-            input: Value::Null,
-            timestamp: Utc::now(),
-        }];
-
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(run_workflow_strict(
-                exec_id,
-                history,
-                echo_workflow,
-                Value::Null,
-                empty_shared_state(),
-            ));
-
-        // 1. Span must be named correctly.
-        let span_fields = fields.fields_for("harvest.workflow.execute");
-        assert!(
-            !span_fields.is_empty(),
-            "run_workflow_strict must emit 'harvest.workflow.execute'; only saw: {:?}",
-            fields
-                .0
-                .lock()
-                .unwrap()
-                .iter()
-                .map(|(n, _)| n.as_str())
-                .collect::<Vec<_>>()
-        );
-
-        // 2. harvest.replay must be `true`.
-        let any_has_replay_true = span_fields.iter().any(|field_set| {
-            field_set
-                .iter()
-                .any(|(name, value)| name == ATTR_REPLAY && value == "true")
-        });
-        assert!(
-            any_has_replay_true,
-            "replay span must carry '{ATTR_REPLAY} = true'; saw: {span_fields:?}"
-        );
-    }
-
-    /// `run_workflow` (live path) must emit `harvest.workflow.execute`
-    /// with `harvest.replay = false`.
-    ///
-    /// RED: current code does not set any `harvest.replay` attribute.
-    #[test]
-    fn live_executor_span_has_replay_false() {
-        use crate::telemetry::ATTR_REPLAY;
-        use span_capture::{SpanFieldLayer, SpanFields};
-        use tracing_subscriber::prelude::*;
-
-        let fields = SpanFields::default();
-        let subscriber = tracing_subscriber::registry().with(SpanFieldLayer(fields.clone()));
-        let _guard = tracing::subscriber::set_default(subscriber);
-
-        let exec_id = ExecutionId::new();
-        let history = vec![WorkflowEvent::WorkflowStarted {
-            input: Value::Null,
-            timestamp: Utc::now(),
-        }];
-
-        tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(run_workflow(exec_id, history, echo_workflow, Value::Null));
-
-        let span_fields = fields.fields_for("harvest.workflow.execute");
-        let any_has_replay_false = span_fields.iter().any(|field_set| {
-            field_set
-                .iter()
-                .any(|(name, value)| name == ATTR_REPLAY && value == "false")
-        });
-        assert!(
-            any_has_replay_false,
-            "live executor span must carry '{ATTR_REPLAY} = false'; saw: {span_fields:?}"
-        );
-    }
 }
